Improved build setup
