+++ common/build.gradle
@@ -14,17 +14,14 @@ dependencies {
 }
 
 shadowJar {
-    dependencies {}
+    dependencies {
+        include(dependency("${group_id}:${project_id}:${version}"))
+        include(dependency("gs.mclo:api:4.0.0"))
+    }
+    relocate("gs.mclo", "${group_id}.${project_id}.libs.mclogs")
 
     destinationDirectory = file('../target')
     archiveFileName = "${project_name}-common-${version}.jar"
-}
-shadowJar.doLast {
-    task ->
-        ant.checksum(file: task.archivePath) {
-            hashAlgorithm = 'md5'
-            property = 'md5sum'
-        }
 }
 
 artifacts {
