+++ platforms/bukkit/build.gradle
@@ -38,13 +38,6 @@ shadowJar {
     destinationDirectory = file('../../target')
     archiveFileName = "${project_name}-bukkit-${minecraft_version}-${version}.jar"
 }
-shadowJar.doLast {
-    task ->
-        ant.checksum(file: task.archivePath) {
-            hashAlgorithm = 'md5'
-            property = 'md5sum'
-        }
-}
 
 artifacts {
     archives(shadowJar)
