+++ platforms/bungee/build.gradle
@@ -36,13 +36,6 @@ shadowJar {
     destinationDirectory = file('../../target')
     archiveFileName = "${project_name}-bungee-${minecraft_version}-${version}.jar"
 }
-shadowJar.doLast {
-    task ->
-        ant.checksum(file: task.archivePath) {
-            hashAlgorithm = 'md5'
-            property = 'md5sum'
-        }
-}
 
 artifacts {
     archives(shadowJar)
