--- platforms/fabric-1.20.2/src/main/resources/taterutils.mixins.json
@@ -1,10 +1,0 @@
-{
-  "required": true,
-  "minVersion": "0.8",
-  "package": "dev.neuralnexus.taterutils.fabric.mixin",
-  "compatibilityLevel": "JAVA_8",
-  "mixins": [],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
