+++ platforms/fabric/gradle.properties
