+++ platforms/fabric/src/main/java/dev/neuralnexus/taterutils/platforms/FabricPlugin.java
@@ -1,0 +1,17 @@
+package dev.neuralnexus.taterutils.platforms;
+
+import dev.neuralnexus.taterlib.logger.LoggerAdapter;
+import dev.neuralnexus.taterutils.TaterUtils;
+import dev.neuralnexus.taterutils.TaterUtilsPlugin;
+
+import net.fabricmc.api.ModInitializer;
+
+/** Fabric entry point. */
+public class FabricPlugin implements ModInitializer, TaterUtilsPlugin {
+    public FabricPlugin() {
+        pluginStart(this, new LoggerAdapter(TaterUtils.Constants.PROJECT_NAME));
+    }
+
+    @Override
+    public void onInitialize() {}
+}
