+++ platforms/fabric/src/main/resources/assets/TaterUtils/icon.png
