+++ platforms/fabric/src/main/resources/fabric.mod.json
@@ -1,0 +1,38 @@
+{
+  "schemaVersion": 1,
+  "id": "${project_id}",
+  "version": "${version}",
+  "name": "${project_name}",
+  "description": "${project_description}",
+  "authors": [
+    {
+      "name": "p0t4t0sandwich",
+      "contact": {
+        "sources": "https://github.com/p0t4t0sandwich/",
+        "homepage": "https://links.sperrer.ca/"
+      }
+    }
+  ],
+  "contact": {
+    "repo": "${repo_url}"
+  },
+  "license": "${license}",
+  "icon": "assets/${project_name}/icon.png",
+  "environment": "*",
+  "entrypoints": {
+    "main": [
+      "${group_id}.${project_id}.platforms.FabricPlugin"
+    ]
+  },
+  "mixins": [
+    "${project_id}.mixins.json"
+  ],
+  "depends": {
+    "fabricloader": ">=0.9.0",
+    "minecraft": ">=1.14",
+    "taterlib": ">=${taterlib_version}"
+  },
+  "suggests": {
+    "luckperms": "*"
+  }
+}
