--- platforms/forge-1.13.2/src/main/resources/META-INF/mods.toml
@@ -1,41 +1,0 @@
-# https://files.minecraftforge.net/net/minecraftforge/forge/
-modLoader = "javafml"
-loaderVersion = "${loader_version_range}"
-license = "${license}"
-issueTrackerURL = "${issue_tracker_url}"
-
-[[mods]]
-modId = "${project_id}"
-version = "${version}"
-displayName = "${project_name}"
-updateJSONURL = "${update_json_url}"
-displayURL = "${homepage_url}"
-logoFile = "${project_name}.png"
-credits = "${authors}"
-authors = "${authors}"
-displayTest = "IGNORE_SERVER_VERSION"
-description = '''${project_description}'''
-
-# Forge Dependency
-[[dependencies.${ project_id }]]
-modId = "forge"
-mandatory = true
-versionRange = "${forge_version_range}"
-ordering = "NONE"
-side = "BOTH"
-
-# Minecraft Dependency
-[[dependencies.${ project_id }]]
-modId = "minecraft"
-mandatory = true
-versionRange = "${minecraft_version_range}"
-ordering = "NONE"
-side = "BOTH"
-
-# LuckPerms Dependency
-[[dependencies.${ project_id }]]
-modId = "luckperms"
-mandatory = false
-versionRange = "[5.3,)"
-ordering = "NONE"
-side = "SERVER"
