+++ platforms/forge/gradle.properties
