+++ platforms/forge/src/main/resources/META-INF/mods.toml
@@ -1,0 +1,49 @@
+# https://files.minecraftforge.net/net/minecraftforge/forge/
+modLoader = "javafml"
+loaderVersion = "${loader_version_range}"
+license = "${license}"
+issueTrackerURL = "${issue_tracker_url}"
+
+[[mods]]
+modId = "${project_id}"
+version = "${version}"
+displayName = "${project_name}"
+updateJSONURL = "${update_json_url}"
+displayURL = "${homepage_url}"
+logoFile = "${project_name}.png"
+credits = "${authors}"
+authors = "${authors}"
+displayTest = "IGNORE_SERVER_VERSION"
+description = '''${project_description}'''
+
+# Forge Dependency
+[[dependencies.${project_id}]]
+modId = "forge"
+mandatory = true
+versionRange = "${forge_version_range}"
+ordering = "NONE"
+side = "BOTH"
+
+# Minecraft Dependency
+[[dependencies.${project_id}]]
+modId = "minecraft"
+mandatory = true
+versionRange = "${minecraft_version_range}"
+ordering = "NONE"
+side = "BOTH"
+
+# TaterLib Dependency
+[[dependencies.${project_id}]]
+modId = "taterlib"
+mandatory = true
+versionRange = "[${taterlib_version},)"
+ordering = "NONE"
+side = "SERVER"
+
+# LuckPerms Dependency
+[[dependencies.${project_id}]]
+modId = "luckperms"
+mandatory = false
+versionRange = "[5.3,)"
+ordering = "NONE"
+side = "SERVER"
