+++ platforms/neoforge/gradle.properties
