+++ platforms/sponge/gradle.properties
