--- platforms/sponge8/build.gradle
@@ -1,69 +1,0 @@
-import org.spongepowered.gradle.plugin.config.PluginLoaders
-import org.spongepowered.plugin.metadata.model.PluginDependency
-
-plugins {
-    id('com.github.johnrengelman.shadow') version('8.1.1')
-    id('java-library')
-    id("org.spongepowered.gradle.plugin") version('2.1.1')
-}
-
-sponge {
-    apiVersion(spongeapi_version)
-    license("GPL-3.0")
-    loader {
-        name(PluginLoaders.JAVA_PLAIN)
-        version("1.0")
-    }
-    plugin(project_id) {
-        version(version)
-        displayName(project_name)
-        entrypoint("${group_id}.${project_id}.sponge.SpongePlugin")
-        description(project_description)
-        links {
-            homepage(homepage_url)
-            source(repo_url)
-            issues(issue_tracker_url)
-        }
-        dependency("spongeapi") {
-            loadOrder(PluginDependency.LoadOrder.AFTER)
-            optional(false)
-        }
-        dependency("taterlib") {
-            version(taterlib_version)
-            loadOrder(PluginDependency.LoadOrder.AFTER)
-            optional(false)
-        }
-        dependency("luckperms") {
-            version("5.3.0")
-            loadOrder(PluginDependency.LoadOrder.AFTER)
-            optional(true)
-        }
-    }
-}
-
-dependencies {
-    implementation(project(path: ':common', configuration: 'shadow'))
-
-    // TaterLib
-    compileOnly("dev.neuralnexus:TaterLib:${taterlib_version}")
-}
-
-shadowJar {
-    dependencies {
-        include(project(":common"))
-    }
-
-    destinationDirectory = file('../../target')
-    archiveFileName = "${project_name}-sponge8-${version}.jar"
-}
-shadowJar.doLast {
-    task ->
-        ant.checksum(file: task.archivePath) {
-            hashAlgorithm = 'md5'
-            property = 'md5sum'
-        }
-}
-
-artifacts {
-    archives(shadowJar)
-}
