+++ platforms/velocity/build.gradle
@@ -1,0 +1,26 @@
+plugins {
+    id('com.github.johnrengelman.shadow') version('8.1.1')
+}
+
+dependencies {
+    implementation(project(path: ':common', configuration: 'shadow'))
+
+    compileOnly("com.velocitypowered:velocity-api:${velocityapi_version}")
+    annotationProcessor("com.velocitypowered:velocity-api:${velocityapi_version}")
+
+    // TaterLib
+    compileOnly("dev.neuralnexus:TaterLib:${taterlib_version}")
+}
+
+shadowJar {
+    dependencies {
+        include(project(":common"))
+    }
+
+    destinationDirectory = file('../../target')
+    archiveFileName = "${project_name}-velocity3-${version}.jar"
+}
+
+artifacts {
+    archives(shadowJar)
+}
