+++ platforms/velocity/gradle.properties
