--- platforms/velocity3/build.gradle
@@ -1,33 +1,0 @@
-plugins {
-    id('com.github.johnrengelman.shadow') version('8.1.1')
-}
-
-dependencies {
-    implementation(project(path: ':common', configuration: 'shadow'))
-
-    compileOnly("com.velocitypowered:velocity-api:${velocityapi_version}")
-    annotationProcessor("com.velocitypowered:velocity-api:${velocityapi_version}")
-
-    // TaterLib
-    compileOnly("dev.neuralnexus:TaterLib:${taterlib_version}")
-}
-
-shadowJar {
-    dependencies {
-        include(project(":common"))
-    }
-
-    destinationDirectory = file('../../target')
-    archiveFileName = "${project_name}-velocity3-${version}.jar"
-}
-shadowJar.doLast {
-    task ->
-        ant.checksum(file: task.archivePath) {
-            hashAlgorithm = 'md5'
-            property = 'md5sum'
-        }
-}
-
-artifacts {
-    archives(shadowJar)
-}
